third_party/blink/web_tests/platform/generic/fast/dom/gc-custom-prototype-expected.txt

Tests that custom prototypes on DOM objects persist after garbage collection.

PASS: $('p').__proto__ should be null and is.
PASS: $('p').__proto__ should be null and is.
PASS: $('p').__proto__ should be null and is.
